refactor(D-166): kill the God-file — chat_with_agent + fallbacks + socratic eval into mixins
Slices 4-6 (verbatim moves, D-164 pattern — one manifest line each):
- Slice 4: LocalFallbackMixin (local_fallback.py, ~460 lines) — file-count /
  indexed+DB retrieval / explanation streaming / question-only / local graph
  / general-chat responders.
- Slice 5: SocraticEvaluationMixin (socratic_evaluation.py, ~205 lines).
- Slice 6: ChatTurnMixin (chat_turn.py, ~1,850 lines) — chat_with_agent, the
  last God-method, plus the EXPLANATION_VIA_ORCHESTRATOR rollback lever.

orchestrator_client.py: 2,786 -> 267 lines (-90% this round; -95.7%
cumulative since the 6,154-line God-file). The client is now a thin
transport shell (decisions + JWT + missions). Brain stays first base in
the MRO; behaviour is byte-identical.

Lockstep sweep (alignment, not disablement — D-105 rule 9): 19 test files
+ 7 fitness-gate readers converted from direct client reads to the D-164
manifest (read_tutor_source) — including the D-117 negative anti-prepend
guard which must scan the composed source to keep its value.
