+++ app/infrastructure/clients/orchestrator/tutor_sources.py
@@ -28,6 +28,9 @@ TUTOR_SOURCE_FILES: tuple[str, ...] = (
     "app/infrastructure/clients/orchestrator/stream_normalization.py",  # D-164 Slice 1
     "app/infrastructure/clients/orchestrator/text_streaming.py",  # D-164 Slice 2
     "app/infrastructure/clients/orchestrator/probability_ui.py",  # D-164 Slice 3
+    "app/infrastructure/clients/orchestrator/local_fallback.py",  # D-166 Slice 4
+    "app/infrastructure/clients/orchestrator/socratic_evaluation.py",  # D-166 Slice 5
+    "app/infrastructure/clients/orchestrator/chat_turn.py",  # D-166 Slice 6
 )
